requirements.txt
fastapi
uvicorn
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.64
rich
python-dotenv
tiktoken
typing-extensions
pydantic